feat: refactor DOCX utilities and update dependencies for improved processing
