+++ mineru/model/docx/tools/__init__.py
@@ -1,0 +1,1 @@
+# Copyright (c) Opendatalab. All rights reserved.
