--- mineru/model/utils/docx/pre_process.py
@@ -1,169 +1,0 @@
-import zipfile
-from io import BytesIO
-from typing import BinaryIO
-from xml.etree import ElementTree as ET
-
-from bs4 import BeautifulSoup, Tag
-
-from .math.omml import OMML_NS, oMath2Latex
-
-MATH_ROOT_TEMPLATE = "".join(
-    (
-        "<w:document ",
-        'xmlns:wpc="http://schemas.microsoft.com/office/word/2010/wordprocessingCanvas" ',
-        'xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" ',
-        'xmlns:o="urn:schemas-microsoft-com:office:office" ',
-        'xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" ',
-        'xmlns:m="http://schemas.openxmlformats.org/officeDocument/2006/math" ',
-        'xmlns:v="urn:schemas-microsoft-com:vml" ',
-        'xmlns:wp14="http://schemas.microsoft.com/office/word/2010/wordprocessingDrawing" ',
-        'xmlns:wp="http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing" ',
-        'xmlns:w10="urn:schemas-microsoft-com:office:word" ',
-        'xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main" ',
-        'xmlns:w14="http://schemas.microsoft.com/office/word/2010/wordml" ',
-        'xmlns:wpg="http://schemas.microsoft.com/office/word/2010/wordprocessingGroup" ',
-        'xmlns:wpi="http://schemas.microsoft.com/office/word/2010/wordprocessingInk" ',
-        'xmlns:wne="http://schemas.microsoft.com/office/word/2006/wordml" ',
-        'xmlns:wps="http://schemas.microsoft.com/office/word/2010/wordprocessingShape" mc:Ignorable="w14 wp14">',
-        "{0}</w:document>",
-    )
-)
-
-
-def _convert_omath_to_latex(tag: Tag) -> str:
-    """
-    Converts an OMML (Office Math Markup Language) tag to LaTeX format.
-
-    Args:
-        tag (Tag): A BeautifulSoup Tag object representing the OMML element.
-
-    Returns:
-        str: The LaTeX representation of the OMML element.
-    """
-    # Format the tag into a complete XML document string
-    math_root = ET.fromstring(MATH_ROOT_TEMPLATE.format(str(tag)))
-    # Find the 'oMath' element within the XML document
-    math_element = math_root.find(OMML_NS + "oMath")
-    # Convert the 'oMath' element to LaTeX using the oMath2Latex function
-    latex = oMath2Latex(math_element).latex
-    return latex
-
-
-def _get_omath_tag_replacement(tag: Tag, block: bool = False) -> Tag:
-    """
-    Creates a replacement tag for an OMML (Office Math Markup Language) element.
-
-    Args:
-        tag (Tag): A BeautifulSoup Tag object representing the "oMath" element.
-        block (bool, optional): If True, the LaTeX will be wrapped in double dollar signs for block mode. Defaults to False.
-
-    Returns:
-        Tag: A BeautifulSoup Tag object representing the replacement element.
-    """
-    t_tag = Tag(name="w:t")
-    t_tag.string = (
-        f"$${_convert_omath_to_latex(tag)}$$"
-        if block
-        else f"${_convert_omath_to_latex(tag)}$"
-    )
-    r_tag = Tag(name="w:r")
-    r_tag.append(t_tag)
-    return r_tag
-
-
-def _replace_equations(tag: Tag):
-    """
-    Replaces OMML (Office Math Markup Language) elements with their LaTeX equivalents.
-
-    Args:
-        tag (Tag): A BeautifulSoup Tag object representing the OMML element. Could be either "oMathPara" or "oMath".
-
-    Raises:
-        ValueError: If the tag is not supported.
-    """
-    if tag.name == "oMathPara":
-        # Create a new paragraph tag
-        p_tag = Tag(name="w:p")
-        # Replace each 'oMath' child tag with its LaTeX equivalent as block equations
-        for child_tag in tag.find_all("oMath"):
-            p_tag.append(_get_omath_tag_replacement(child_tag, block=True))
-        # Replace the original 'oMathPara' tag with the new paragraph tag
-        tag.replace_with(p_tag)
-    elif tag.name == "oMath":
-        # Replace the 'oMath' tag with its LaTeX equivalent as inline equation
-        tag.replace_with(_get_omath_tag_replacement(tag, block=False))
-    else:
-        raise ValueError(f"Not supported tag: {tag.name}")
-
-
-def _pre_process_math(content: bytes) -> bytes:
-    """
-    Pre-processes the math content in a DOCX -> XML file by converting OMML (Office Math Markup Language) elements to LaTeX.
-    This preprocessed content can be directly replaced in the DOCX file -> XMLs.
-
-    Args:
-        content (bytes): The XML content of the DOCX file as bytes.
-
-    Returns:
-        bytes: The processed content with OMML elements replaced by their LaTeX equivalents, encoded as bytes.
-    """
-    soup = BeautifulSoup(content.decode(), features="xml")
-    for tag in soup.find_all("oMathPara"):
-        _replace_equations(tag)
-    for tag in soup.find_all("oMath"):
-        _replace_equations(tag)
-    return str(soup).encode()
-
-
-def pre_process_docx(input_docx: BinaryIO) -> BinaryIO:
-    """
-    预处理DOCX文件，执行指定的处理步骤。
-
-    处理过程包括：在内存中解压DOCX文件，转换特定的XML文件（如将OMML元素转换为LaTeX），
-    然后将所有内容重新打包成一个新的DOCX文件，整个过程不写入磁盘。
-
-    Args:
-        input_docx (BinaryIO): 表示DOCX文件的二进制输入流。
-
-    Returns:
-        BinaryIO: 表示已处理DOCX文件的二进制输出流。
-    """
-    # 创建用于存储输出DOCX文件的BytesIO对象
-    output_docx = BytesIO()
-    
-    # 定义需要预处理的文件列表
-    pre_process_enable_files = [
-        "word/document.xml",      # 主文档内容
-        "word/footnotes.xml",     # 脚注
-        "word/endnotes.xml",      # 尾注
-    ]
-    
-    # 读取输入的DOCX文件（ZIP格式）
-    with zipfile.ZipFile(input_docx, mode="r") as zip_input:
-        # 提取所有文件到内存字典中
-        files = {name: zip_input.read(name) for name in zip_input.namelist()}
-        
-        # 创建新的ZIP文件用于输出
-        with zipfile.ZipFile(output_docx, mode="w") as zip_output:
-            # 保留原始ZIP文件的注释
-            zip_output.comment = zip_input.comment
-            
-            # 遍历所有文件
-            for name, content in files.items():
-                # 检查文件是否需要预处理
-                if name in pre_process_enable_files:
-                    try:
-                        # 对指定文件执行数学公式预处理（OMML转LaTeX）
-                        updated_content = _pre_process_math(content)
-                        # 将处理后的内容写入输出文件
-                        zip_output.writestr(name, updated_content)
-                    except Exception:
-                        # 如果处理过程中出现异常，保留原始内容
-                        zip_output.writestr(name, content)
-                else:
-                    # 不需要处理的文件直接复制
-                    zip_output.writestr(name, content)
-    
-    # 将文件指针移到开头，准备返回
-    output_docx.seek(0)
-    return output_docx
