+++ pyproject.toml
@@ -43,6 +43,7 @@ dependencies = [
     "qwen-vl-utils>=0.0.14,<1",
     "defusedxml>=0.7.1,<1",
     "python-docx>=1.2.0,<2",
+    "mammoth>=1.11.0,<2",
 ]
 
 [project.optional-dependencies]
